56 Sunningdale Drive
"With making away with Mr. Neville St.--Oh, come, you can't be charged with that unless they make a
case of attempted suicide of it," said the inspector with a grin. "Well, I have been twenty-seven
years in the force, but this really takes the cake."

"If I am Mr. Neville St. Clair, then it is obvious that no crime has been committed, and that,
therefore, I am illegally detained."