32 Harvey Street
With a dazed face the banker made out the required check. Holmes walked over to his desk, took out a
little triangular piece of gold with three gems in it, and threw it down upon the table.

With a shriek of joy our client clutched it up.

"You have it!" he gasped. "I am saved! I am saved!"