31 Beaconside
"Do not think of revenge, or anything of the sort, at present. I think that we may gain that by
means of the law; but we have our web to weave, while theirs is already woven. The first
consideration is to remove the pressing danger which threatens you. The second is to clear up the
mystery and to punish the guilty parties."

"I thank you," said the young man, rising and pulling on his overcoat. "You have given me fresh life
and hope. I shall certainly do as you advise."

"Do not lose an instant. And, above all, take care of yourself in the meanwhile, for I do not think
that there can be a doubt that you are threatened by a very real and imminent danger. How do you go
back?"

"By train from Waterloo."

"It is not yet nine. The streets will be crowded, so I trust that you may be in safety. And yet you
cannot guard yourself too closely."

"I am armed."