95 Larch Walk
"The case has been an interesting one," remarked Holmes when our visitors had left us, "because it
serves to show very clearly how simple the explanation may be of an affair which at first sight
seems to be almost inexplicable. Nothing could be more natural than the sequence of events as
narrated by this lady, and nothing stranger than the result when viewed, for instance, by Mr.
Lestrade of Scotland Yard."

"You were not yourself at fault at all, then?"