88 Moor End
"Very good. Come this way, if you please." He led us down a passage, opened a barred door, passed
down a winding stair, and brought us to a whitewashed corridor with a line of doors on each side.

"The third on the right is his," said the inspector. "Here it is!" He quietly shot back a panel in
the upper part of the door and glanced through.

"He is asleep," said he. "You can see him very well."

We both put our eyes to the grating. The prisoner lay with his face towards us, in a very deep
sleep, breathing slowly and heavily. He was a middle-sized man, coarsely clad as became his calling,
with a coloured shirt protruding through the rent in his tattered coat. He was, as the inspector had
said, extremely dirty, but the grime which covered his face could not conceal its repulsive
ugliness. A broad wheal from an old scar ran right across it from eye to chin, and by its
contraction had turned up one side of the upper lip, so that three teeth were exposed in a perpetual
snarl. A shock of very bright red hair grew low over his eyes and forehead.