128 Wantage Close
"It appears that his arrest did not take place at once, but after the return to Hatherley Farm. On
the inspector of constabulary informing him that he was a prisoner, he remarked that he was not
surprised to hear it, and that it was no more than his deserts. This observation of his had the
natural effect of removing any traces of doubt which might have remained in the minds of the
coroner's jury."

"It was a confession," I ejaculated.

"No, for it was followed by a protestation of innocence."