1 Wantage Close
"What of the rat, then?"

Sherlock Holmes took a folded paper from his pocket and flattened it out on the table. "This is a
map of the Colony of Victoria," he said. "I wired to Bristol for it last night." He put his hand
over part of the map. "What do you read?"