64 Moor End
"It is not yet nine. The streets will be crowded, so I trust that you may be in safety. And yet you
cannot guard yourself too closely."

"I am armed."

"That is well. To-morrow I shall set to work upon your case."

"I shall see you at Horsham, then?"

"No, your secret lies in London. It is there that I shall seek it."

"Then I shall call upon you in a day, or in two days, with news as to the box and the papers. I
shall take your advice in every particular." He shook hands with us and took his leave. Outside the
wind still screamed and the rain splashed and pattered against the windows. This strange, wild story
seemed to have come to us from amid the mad elements--blown in upon us like a sheet of sea-weed in a
gale--and now to have been reabsorbed by them once more.