37 Balcombe Close
"Very glad to see you. I dine at seven. There is a woodcock, I believe. By the way, in view of
recent occurrences, perhaps I ought to ask Mrs. Hudson to examine its crop."

I had been delayed at a case, and it was a little after half-past six when I found myself in Baker
Street once more. As I approached the house I saw a tall man in a Scotch bonnet with a coat which
was buttoned up to his chin waiting outside in the bright semicircle which was thrown from the
fanlight. Just as I arrived the door was opened, and we were shown up together to Holmes' room.