137 Wilson Gardens
"The newcomers were Colonel Lysander Stark and a short thick man with a chinchilla beard growing out
of the creases of his double chin, who was introduced to me as Mr. Ferguson.

"'This is my secretary and manager,' said the colonel. 'By the way, I was under the impression that
I left this door shut just now. I fear that you have felt the draught.'

"'On the contrary,' said I, 'I opened the door myself because I felt the room to be a little close.'

"He shot one of his suspicious looks at me. 'Perhaps we had better proceed to business, then,' said
he. 'Mr. Ferguson and I will take you up to see the machine.'

"'I had better put my hat on, I suppose.'

"'Oh, no, it is in the house.'