5 Valentia Road
"Sarasate plays at the St. James's Hall this afternoon," he remarked. "What do you think, Watson?
Could your patients spare you for a few hours?"

"I have nothing to do to-day. My practice is never very absorbing."

"Then put on your hat and come. I am going through the City first, and we can have some lunch on the
way. I observe that there is a good deal of German music on the programme, which is rather more to
my taste than Italian or French. It is introspective, and I want to introspect. Come along!"

We travelled by the Underground as far as Aldersgate; and a short walk took us to Saxe-Coburg
Square, the scene of the singular story which we had listened to in the morning. It was a poky,
little, shabby-genteel place, where four lines of dingy two-storied brick houses looked out into a
small railed-in enclosure, where a lawn of weedy grass and a few clumps of faded laurel-bushes made
a hard fight against a smoke-laden and uncongenial atmosphere. Three gilt balls and a brown board
with "Jabez Wilson" in white letters, upon a corner house, announced the place where our red-headed
client carried on his business. Sherlock Holmes stopped in front of it with his head on one side and
looked it all over, with his eyes shining brightly between puckered lids. Then he walked slowly up
the street, and then down again to the corner, still looking keenly at the houses. Finally he
returned to the pawnbroker's, and, having thumped vigorously upon the pavement with his stick two or
three times, he went up to the door and knocked. It was instantly opened by a bright-looking, clean-
shaven young fellow, who asked him to step in.