76 Glan Road
There was a movement and an exclamation from my right, and peering through the gloom, I saw Whitney,
pale, haggard, and unkempt, staring out at me.

"My God! It's Watson," said he. He was in a pitiable state of reaction, with every nerve in a
twitter. "I say, Watson, what o'clock is it?"

"Nearly eleven."

"Of what day?"

"Of Friday, June 19th."