42 Valentia Road
"Won't it ring?"

"No, it is not even attached to a wire. This is very interesting. You can see now that it is
fastened to a hook just above where the little opening for the ventilator is."

"How very absurd! I never noticed that before."

"Very strange!" muttered Holmes, pulling at the rope. "There are one or two very singular points
about this room. For example, what a fool a builder must be to open a ventilator into another room,
when, with the same trouble, he might have communicated with the outside air!"