17 Wantage Close
A professional case of great gravity was engaging my own attention at the time, and the whole of
next day I was busy at the bedside of the sufferer. It was not until close upon six o'clock that I
found myself free and was able to spring into a hansom and drive to Baker Street, half afraid that I
might be too late to assist at the dénouement of the little mystery. I found Sherlock Holmes alone,
however, half asleep, with his long, thin form curled up in the recesses of his armchair. A
formidable array of bottles and test-tubes, with the pungent cleanly smell of hydrochloric acid,
told me that he had spent his day in the chemical work which was so dear to him.

"Well, have you solved it?" I asked as I entered.

"Yes. It was the bisulphate of baryta."

"No, no, the mystery!" I cried.

"Oh, that! I thought of the salt that I have been working upon. There was never any mystery in the
matter, though, as I said yesterday, some of the details are of interest. The only drawback is that
there is no law, I fear, that can touch the scoundrel."