100 Moor End
"Well, when I came to think it all over in cool blood I was very much astonished, as you may both
think, at this sudden commission which had been intrusted to me. On the one hand, of course, I was
glad, for the fee was at least tenfold what I should have asked had I set a price upon my own
services, and it was possible that this order might lead to other ones. On the other hand, the face
and manner of my patron had made an unpleasant impression upon me, and I could not think that his
explanation of the fuller's-earth was sufficient to explain the necessity for my coming at midnight,
and his extreme anxiety lest I should tell anyone of my errand. However, I threw all fears to the
winds, ate a hearty supper, drove to Paddington, and started off, having obeyed to the letter the
injunction as to holding my tongue.

"At Reading I had to change not only my carriage but my station. However, I was in time for the last
train to Eyford, and I reached the little dim-lit station after eleven o'clock. I was the only
passenger who got out there, and there was no one upon the platform save a single sleepy porter with
a lantern. As I passed out through the wicket gate, however, I found my acquaintance of the morning
waiting in the shadow upon the other side. Without a word he grasped my arm and hurried me into a
carriage, the door of which was standing open. He drew up the windows on either side, tapped on the
wood-work, and away we went as fast as the horse could go."

"One horse?" interjected Holmes.

"Yes, only one."

"Did you observe the colour?"