63 Moor End
It was late before Sherlock Holmes returned. He came back alone, for Lestrade was staying in
lodgings in the town.

"The glass still keeps very high," he remarked as he sat down. "It is of importance that it should
not rain before we are able to go over the ground. On the other hand, a man should be at his very
best and keenest for such nice work as that, and I did not wish to do it when fagged by a long
journey. I have seen young McCarthy."

"And what did you learn from him?"

"Nothing."

"Could he throw no light?"

"None at all. I was inclined to think at one time that he knew who had done it and was screening him
or her, but I am convinced now that he is as puzzled as everyone else. He is not a very quick-witted
youth, though comely to look at and, I should think, sound at heart."