92 Badgers Dene
"Not only that, but the signature is typewritten. Look at the neat little 'Hosmer Angel' at the
bottom. There is a date, you see, but no superscription except Leadenhall Street, which is rather
vague. The point about the signature is very suggestive--in fact, we may call it conclusive."

"Of what?"

"My dear fellow, is it possible you do not see how strongly it bears upon the case?"

"I cannot say that I do unless it were that he wished to be able to deny his signature if an action
for breach of promise were instituted."

"No, that was not the point. However, I shall write two letters, which should settle the matter. One
is to a firm in the City, the other is to the young lady's stepfather, Mr. Windibank, asking him
whether he could meet us here at six o'clock tomorrow evening. It is just as well that we should do
business with the male relatives. And now, Doctor, we can do nothing until the answers to those
letters come, so we may put our little problem upon the shelf for the interim."

I had had so many reasons to believe in my friend's subtle powers of reasoning and extraordinary
energy in action that I felt that he must have some solid grounds for the assured and easy demeanour
with which he treated the singular mystery which he had been called upon to fathom. Once only had I
known him to fail, in the case of the King of Bohemia and of the Irene Adler photograph; but when I
looked back to the weird business of the Sign of Four, and the extraordinary circumstances connected
with the Study in Scarlet, I felt that it would be a strange tangle indeed which he could not
unravel.