101 Holmefield Avenue
"But I am somewhat headstrong by nature, and the more ready to engage in an affair when there is
some obstacle in the way. I thought of my fifty-guinea fee, of my wearisome journey, and of the
unpleasant night which seemed to be before me. Was it all to go for nothing? Why should I slink away
without having carried out my commission, and without the payment which was my due? This woman
might, for all I knew, be a monomaniac. With a stout bearing, therefore, though her manner had
shaken me more than I cared to confess, I still shook my head and declared my intention of remaining
where I was. She was about to renew her entreaties when a door slammed overhead, and the sound of
several footsteps was heard upon the stairs. She listened for an instant, threw up her hands with a
despairing gesture, and vanished as suddenly and as noiselessly as she had come.

"The newcomers were Colonel Lysander Stark and a short thick man with a chinchilla beard growing out
of the creases of his double chin, who was introduced to me as Mr. Ferguson.

"'This is my secretary and manager,' said the colonel. 'By the way, I was under the impression that
I left this door shut just now. I fear that you have felt the draught.'

"'On the contrary,' said I, 'I opened the door myself because I felt the room to be a little close.'

"He shot one of his suspicious looks at me. 'Perhaps we had better proceed to business, then,' said
he. 'Mr. Ferguson and I will take you up to see the machine.'

"'I had better put my hat on, I suppose.'