15 Wantage Close
"Ay, bodies, Watson. We should be rich men if we had £1000 for every poor devil who has been done to
death in that den. It is the vilest murder-trap on the whole riverside, and I fear that Neville St.
Clair has entered it never to leave it more. But our trap should be here." He put his two
forefingers between his teeth and whistled shrilly--a signal which was answered by a similar whistle
from the distance, followed shortly by the rattle of wheels and the clink of horses' hoofs.

"Now, Watson," said Holmes, as a tall dog-cart dashed up through the gloom, throwing out two golden
tunnels of yellow light from its side lanterns. "You'll come with me, won't you?"

"If I can be of use."

"Oh, a trusty comrade is always of use; and a chronicler still more so. My room at The Cedars is a
double-bedded one."

"The Cedars?"