150 Valentia Road
"They are typewritten," I remarked.

"Not only that, but the signature is typewritten. Look at the neat little 'Hosmer Angel' at the
bottom. There is a date, you see, but no superscription except Leadenhall Street, which is rather
vague. The point about the signature is very suggestive--in fact, we may call it conclusive."