47 Badgers Dene
"But if he is innocent, who has done it?"

"Ah! who? I would call your attention very particularly to two points. One is that the murdered man
had an appointment with someone at the pool, and that the someone could not have been his son, for
his son was away, and he did not know when he would return. The second is that the murdered man was
heard to cry 'Cooee!' before he knew that his son had returned. Those are the crucial points upon
which the case depends. And now let us talk about George Meredith, if you please, and we shall leave
all minor matters until to-morrow."