95 Bowler Street
"Last week he hurled the local blacksmith over a parapet into a stream, and it was only by paying
over all the money which I could gather together that I was able to avert another public exposure.
He had no friends at all save the wandering gypsies, and he would give these vagabonds leave to
encamp upon the few acres of bramble-covered land which represent the family estate, and would
accept in return the hospitality of their tents, wandering away with them sometimes for weeks on
end. He has a passion also for Indian animals, which are sent over to him by a correspondent, and he
has at this moment a cheetah and a baboon, which wander freely over his grounds and are feared by
the villagers almost as much as their master.

"You can imagine from what I say that my poor sister Julia and I had no great pleasure in our lives.
No servant would stay with us, and for a long time we did all the work of the house. She was but
thirty at the time of her death, and yet her hair had already begun to whiten, even as mine has."