87 Moor End
"You have an answer to everything. But how on earth do you deduce that the gas is not laid on in his
house?"

"One tallow stain, or even two, might come by chance; but when I see no less than five, I think that
there can be little doubt that the individual must be brought into frequent contact with burning
tallow--walks upstairs at night probably with his hat in one hand and a guttering candle in the
other. Anyhow, he never got tallow-stains from a gas-jet. Are you satisfied?"